+++ app/layout.tsx
@@ -1,4 +1,3 @@
-import "react-toastify/dist/ReactToastify.min.css";
 import "@/styles/globals.css";
 
 import type { Metadata, Viewport } from "next";
