+++ package.json
@@ -17,7 +17,6 @@
   },
   "dependencies": {
     "@heroicons/react": "^2.2.0",
-    "@radix-ui/react-tooltip": "^1.0.5",
     "class-variance-authority": "^0.6.0",
     "classnames": "^2.3.2",
     "cmdk": "^1.1.1",
@@ -27,7 +26,6 @@
     "openai": "^4.96.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-toastify": "^9.1.1",
     "tailwind-merge": "^1.10.0",
     "tailwindcss-animate": "^1.0.5",
     "typescript": "5.1.6"
