+++ styles/globals.css
@@ -4,81 +4,84 @@
 
 @layer base {
   :root {
-    /* Light: warm off-white paper */
-    --background: 45 18% 97%;
-    --foreground: 30 8% 12%;
+    /* OKLCH: L C H. Light: warm off-white paper, warm-brown ink. */
+    --background: 0.972 0.012 86;
+    --foreground: 0.215 0.008 60;
 
-    --muted: 40 6% 94%;
-    --muted-foreground: 30 5% 40%;
+    --muted: 0.945 0.008 80;
+    --muted-foreground: 0.46 0.01 60;
 
-    --popover: 45 14% 99%;
-    --popover-foreground: 30 8% 12%;
+    --popover: 0.99 0.01 86;
+    --popover-foreground: 0.215 0.008 60;
 
-    --card: 45 14% 99%;
-    --card-foreground: 30 8% 12%;
+    --card: 0.99 0.01 86;
+    --card-foreground: 0.215 0.008 60;
 
-    --border: 40 6% 86%;
-    --input: 40 6% 86%;
+    --border: 0.87 0.008 80;
+    --input: 0.87 0.008 80;
 
-    --primary: 30 8% 12%;
-    --primary-foreground: 45 18% 97%;
+    --primary: 0.215 0.008 60;
+    --primary-foreground: 0.972 0.012 86;
 
-    --secondary: 40 6% 94%;
-    --secondary-foreground: 30 8% 12%;
+    --secondary: 0.945 0.008 80;
+    --secondary-foreground: 0.215 0.008 60;
 
-    --accent: 40 6% 94%;
-    --accent-foreground: 30 8% 12%;
+    --accent: 0.945 0.008 80;
+    --accent-foreground: 0.215 0.008 60;
 
-    --destructive: 0 70% 42%;
-    --destructive-foreground: 45 18% 97%;
+    --destructive: 0.5 0.18 28;
+    --destructive-foreground: 0.972 0.012 86;
 
-    --ring: 30 8% 12%;
+    --ring: 0.215 0.008 60;
 
     /* Brand accent: warm sienna, ties to the cathode-amber voice */
-    --brand: 18 75% 48%;
-    --brand-muted: 18 50% 42%;
+    --brand: 0.58 0.17 45;
+    --brand-muted: 0.52 0.12 45;
 
     /* Available-status indicator: muted, theme-aware */
-    --status: 150 45% 38%;
+    --status: 0.5 0.11 155;
+
+    /* Surface for logos that need a light backdrop in both themes */
+    --logo-surface: 1 0 0;
 
     --radius: 0.375rem;
   }
 
   .dark {
     /* Dark: warm-tinted near-black, off-white text */
-    --background: 30 6% 7%;
-    --foreground: 45 18% 92%;
+    --background: 0.155 0.006 60;
+    --foreground: 0.92 0.015 86;
 
-    --muted: 30 4% 13%;
-    --muted-foreground: 35 6% 62%;
+    --muted: 0.21 0.005 60;
+    --muted-foreground: 0.66 0.008 70;
 
-    --popover: 30 6% 9%;
-    --popover-foreground: 45 18% 92%;
+    --popover: 0.18 0.006 60;
+    --popover-foreground: 0.92 0.015 86;
 
-    --card: 30 6% 9%;
-    --card-foreground: 45 18% 92%;
+    --card: 0.18 0.006 60;
+    --card-foreground: 0.92 0.015 86;
 
-    --border: 30 4% 18%;
-    --input: 30 4% 18%;
+    --border: 0.26 0.005 60;
+    --input: 0.26 0.005 60;
 
-    --primary: 45 18% 92%;
-    --primary-foreground: 30 6% 7%;
+    --primary: 0.92 0.015 86;
+    --primary-foreground: 0.155 0.006 60;
 
-    --secondary: 30 4% 13%;
-    --secondary-foreground: 45 18% 92%;
+    --secondary: 0.21 0.005 60;
+    --secondary-foreground: 0.92 0.015 86;
 
-    --accent: 30 4% 13%;
-    --accent-foreground: 45 18% 92%;
+    --accent: 0.21 0.005 60;
+    --accent-foreground: 0.92 0.015 86;
 
-    --destructive: 0 65% 56%;
-    --destructive-foreground: 45 18% 92%;
+    --destructive: 0.62 0.18 28;
+    --destructive-foreground: 0.92 0.015 86;
 
-    --ring: 45 18% 92%;
+    --ring: 0.92 0.015 86;
 
-    --brand: 25 85% 60%;
-    --brand-muted: 25 55% 52%;
+    --brand: 0.72 0.165 50;
+    --brand-muted: 0.62 0.12 50;
 
-    --status: 150 50% 56%;
+    --status: 0.7 0.13 155;
   }
 }
 
@@ -110,22 +113,11 @@
   body {
     @apply relative bg-background font-mono text-foreground;
     font-feature-settings: "rlig" 1, "calt" 1, "ss01" 1;
+    background-image: url("data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' width='220' height='220'><filter id='n'><feTurbulence type='fractalNoise' baseFrequency='0.9' numOctaves='2' stitchTiles='stitch'/><feColorMatrix type='matrix' values='0 0 0 0 0  0 0 0 0 0  0 0 0 0 0  0 0 0 0.04 0'/></filter><rect width='100%25' height='100%25' filter='url(%23n)'/></svg>");
   }
 
-  body::before {
-    content: "";
-    position: fixed;
-    inset: 0;
-    z-index: 0;
-    pointer-events: none;
-    opacity: 0.045;
-    background-image: url("data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' width='220' height='220'><filter id='n'><feTurbulence type='fractalNoise' baseFrequency='0.9' numOctaves='2' stitchTiles='stitch'/><feColorMatrix type='matrix' values='0 0 0 0 0  0 0 0 0 0  0 0 0 0 0  0 0 0 0.7 0'/></filter><rect width='100%25' height='100%25' filter='url(%23n)'/></svg>");
-    mix-blend-mode: multiply;
-  }
-
-  .dark body::before {
-    opacity: 0.08;
-    mix-blend-mode: screen;
+  .dark body {
+    background-image: url("data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' width='220' height='220'><filter id='n'><feTurbulence type='fractalNoise' baseFrequency='0.9' numOctaves='2' stitchTiles='stitch'/><feColorMatrix type='matrix' values='0 0 0 0 1  0 0 0 0 1  0 0 0 0 1  0 0 0 0.06 0'/></filter><rect width='100%25' height='100%25' filter='url(%23n)'/></svg>");
   }
 
   h1,
